@@ -3,7 +3,7 @@ package com.tistory.deque.previewmaker.kotlin.customview
 import android.content.Context
 import android.graphics.*
 import android.os.AsyncTask
-import android.support.v4.content.ContextCompat
+import androidx.core.content.ContextCompat
 import android.util.AttributeSet
 import android.view.MotionEvent
 import android.view.View
