@@ -1,6 +1,6 @@
 package com.tistory.deque.previewmaker.kotlin.helpmain
 
-import android.support.v7.app.AppCompatActivity
+import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.view.MenuItem
 import com.tistory.deque.previewmaker.R
