@@ -1,14 +1,14 @@
 package com.tistory.deque.previewmaker.kotlin.previewedit
 
 import android.app.AlertDialog
-import android.arch.lifecycle.LiveData
+import androidx.lifecycle.LiveData
 import android.content.Context
 import android.database.CursorIndexOutOfBoundsException
 import android.media.ExifInterface
 import android.net.Uri
 import android.os.AsyncTask
 import android.provider.MediaStore
-import android.support.v4.content.ContextCompat
+import androidx.core.content.ContextCompat
 import com.tistory.deque.previewmaker.R
 import com.tistory.deque.previewmaker.kotlin.base.BaseKotlinViewModel
 import com.tistory.deque.previewmaker.kotlin.db.KtDbOpenHelper
