@@ -1,7 +1,7 @@
 package com.tistory.deque.previewmaker.kotlin.makestamp
 
-import android.arch.lifecycle.LiveData
-import android.arch.lifecycle.MutableLiveData
+import androidx.lifecycle.LiveData
+import androidx.lifecycle.MutableLiveData
 import android.content.ContentResolver
 import android.content.Context
 import android.content.Intent
